+++ flake.nix
@@ -13,7 +13,7 @@
     {
       devShells.${system}.default = pkgs.mkShell {
         buildInputs = with pkgs; [
-          go_1_25
+          go_1_26
           gopls
           delve
           go-task